enable filetype detection for indentation and plugins
This also has the pleasant side effect of detecting when writing a git
commit message and having it automatically wrap at 72 characters.
Although going back to edit seems to undo that functionality for that
specific line, allowing lines that are longer than 72 characters. I'll
try this for now and see how I like it.