set backupdir and undodir to keep workspaces clean
When backup and undo are set, some files are created to store that
information for the file.

  .filename.un~
  filename~

These very frequently get in the way, especially because they are
created alongside the file edited and working inside of a git directory.
These changes redirect where these files are managed.

Additionally, the script will make the directories if they don't already
exist with the -p flag.